LayoutTests/platform/ios-sim-deprecated/fast/dom/NavigatorContentUtils/is-protocol-handler-registered-expected.txt

CONSOLE MESSAGE: line 16: ReferenceError: Can't find variable: debug
This test makes sure that navigator.isProtocolHandlerRegistered throws the proper exceptions and returns the default state of handler.
